+++ samplesource.h
@@ -33,7 +33,7 @@ public:
     virtual std::unique_ptr<T[]> getSamples(size_t start, size_t length) = 0;
     virtual void invalidateEvent() { };
     virtual size_t count() = 0;
-    virtual size_t rate() = 0;
+    virtual double rate() = 0;
     virtual float relativeBandwidth() = 0;
     std::type_index sampleType() override;
 };
